feat: as-mocha module for IDE unit test runner support
This adds a new `as-mocha` module and build steps.  When built, it looks
and acts enough like the mocha CLI to allow IDE-based unit test runners to
run it, which in turn runs tstyche and acts as a Reporter, to present the
tstyche test results in the IDE UI.

(For now, this has only been tested with JetBrains/IntelliJ/WebStorm.)

To support this, several of the core `*Result` types got minor refactors
to give them all a concept of parent/child hierarchy.